Shadow acne solved with shadow bias

+++ src/math/common.c
@@ -6,7 +6,7 @@
 /*   By: dda-cunh <dda-cunh@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/15 13:15:44 by dda-cunh          #+#    #+#             */
-/*   Updated: 2024/02/18 12:47:33 by dda-cunh         ###   ########.fr       */
+/*   Updated: 2024/02/18 13:26:45 by dda-cunh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 double	pythagorean_theorem(double a, double b)
 {
-	return (sqrt(pow(a, 2) + pow(b, 2)));
+	return (sqrt(powf(a, 2) + powf(b, 2)));
 }
 
 double	quadratic_smallest_pos(double a, double b, double c)
@@ -23,7 +23,7 @@ double	quadratic_smallest_pos(double a, double b, double c)
 	double	sol1;
 	double	sol2;
 
-	discriminant = pow(b, 2) - 4 * a * c;
+	discriminant = powf(b, 2) - 4 * a * c;
 	if (discriminant >= 0)
 	{
 		sol1 = (-b + sqrt(discriminant)) / (2 * a);
