+++ src/math/point3.c
@@ -6,7 +6,7 @@
 /*   By: dda-cunh <dda-cunh@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/15 13:17:36 by dda-cunh          #+#    #+#             */
-/*   Updated: 2024/02/18 12:47:33 by dda-cunh         ###   ########.fr       */
+/*   Updated: 2024/02/18 13:27:34 by dda-cunh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,9 @@ t_point3	point3_plus_vec3(t_point3 point, t_vec3 vector)
 
 bool	point3_inside_sphere(t_point3 point, t_object_sphere sphere)
 {
-	return (sqrt(pow(point.x - sphere.center.x, 2)
-			+ pow(point.y - sphere.center.y, 2)
-			+ pow(point.z - sphere.center.z, 2)) <= (sphere.diameter / 2));
+	return (sqrt(powf(point.x - sphere.center.x, 2)
+			+ powf(point.y - sphere.center.y, 2)
+			+ powf(point.z - sphere.center.z, 2)) <= (sphere.diameter / 2));
 }
 
 t_vec3	vec3_from_points(t_point3 start, t_point3 end)
@@ -36,6 +36,6 @@ t_vec3	vec3_from_points(t_point3 start, t_point3 end)
 
 double	point3_distance_point3(t_point3 p1, t_point3 p2)
 {
-	return (sqrt(pow((p2.x - p1.x), 2) + pow((p2.y - p1.y), 2)
-			+ pow((p2.z - p1.z), 2)));
+	return (sqrt(powf((p2.x - p1.x), 2) + powf((p2.y - p1.y), 2)
+			+ powf((p2.z - p1.z), 2)));
 }
